Use :b to switch to results buffer in window...
Previously, we opened the results window and the results buffer in one
shot by passing the buffer name to :sp. The problem with this approach
is that :sp takes a *file*, not a buffer. This caused :sp to create a
new results buffer if the working directory had changed (e.g. if
`autochdir` was being used).